Initial Commit: Create file to get Repo info
Need to find a way to get all the information to find the right overlord
to check out. It's late and that sentence I just wrote did not make
sense. Thank you rebasing gods that this commit will be modified. Also
if this commit hasn't been rebased it's because I'm lazy. Also It's
really fucking late. I got work in the morning and Josephine, my wife,
really hates it when I stay up late doing this stuff.